Remove the non-SQL proxy and RBAC stores and move ClientGetter into pkg/client
